use webpack and karma.
- Update travis tests.
- Remove grunt support.
- Update README.
- Add karma support.
- Add webpack support.
- Build browser bundles in `dist` dir.
- Add Node.js 6 support with babel output in `dist` dir.
- Add top level index file to use node6 support as needed.
- Remove old bind and setImmediate test support.
- Move to node and karma test setup files that use a common test file.
