+++ .travis.yml
@@ -7,7 +7,15 @@ sudo: false
 # download test suite and run tests... submodule? meta testing project with
 # all of the reference implementations?
 script:
-  - npm run test-all
+  - if [ "x$BUNDLER" = "x" ]; then npm run test; fi
+  - if [ "x$BUNDLER" != "x" ]; then npm run test-karma; fi
+# only run karma tests for one node version
+matrix:
+  include:
+  - node_js: "6"
+    env: BUNDLER=webpack
+#  - node_js: "6"
+#    env: BUNDLER=browserify
 notifications:
   email:
     on_success: change
