+++ package.json
@@ -1,62 +1,75 @@
 {
   "name": "jsonld-signatures",
   "version": "1.2.2-2",
-  "description": "An implementation of the Linked Data Signatures specification for JSON-LD in node.js.",
-  "homepage": "http://github.com/digitalbazaar/jsonld-signatures",
+  "description": "An implementation of the Linked Data Signatures specifications for JSON-LD in JavaScript.",
+  "homepage": "https://github.com/digitalbazaar/jsonld-signatures",
   "author": {
     "name": "Digital Bazaar, Inc.",
     "email": "support@digitalbazaar.com",
-    "url": "http://digitalbazaar.com/"
+    "url": "https://digitalbazaar.com/"
   },
   "contributors": [
-    {
-      "name": "Dave Longley",
-      "email": "dlongley@digitalbazaar.com"
-    },
-    {
-      "name": "Manu Sporny",
-      "email": "msporny@digitalbazaar.com"
-    },
-    {
-      "name": "David I. Lehn",
-      "email": "dlehn@digitalbazaar.com"
-    }
+    "Dave Longley <dlongley@digitalbazaar.com>",
+    "Manu Sporny <msporny@digitalbazaar.com>",
+    "David I. Lehn <dlehn@digitalbazaar.com>"
   ],
   "repository": {
     "type": "git",
-    "url": "http://github.com/digitalbazaar/jsonld-signatures"
+    "url": "https://github.com/digitalbazaar/jsonld-signatures"
   },
   "bugs": {
     "url": "https://github.com/digitalbazaar/jsonld-signatures/issues",
     "email": "support@digitalbazaar.com"
   },
   "license": "BSD-3-Clause",
-  "main": "lib/jsonld-signatures.js",
+  "main": "lib/index.js",
+  "files": [
+    "browser/*.js",
+    "dist/*.js",
+    "dist/*.js.map",
+    "dist/node6/**/*.js",
+    "lib/*.js",
+    "lib/**/*.js"
+  ],
   "dependencies": {
-    "async": "^1.5.2",
     "bitcore-message": "github:CoMakery/bitcore-message#dist",
-    "commander": "~2.9.0",
-    "es6-promise": "^4.0.5",
     "jsonld": "^0.5.11",
-    "node-forge": "~0.7.1"
+    "node-forge": "^0.7.1"
   },
   "devDependencies": {
-    "chai": "^4.0.2",
-    "grunt": "~1.0.1",
-    "grunt-cli": "~1.2.0",
-    "grunt-contrib-jshint": "~1.1.0",
-    "grunt-jscs": "^3.0.1",
-    "grunt-mocha-test": "~0.13.2",
-    "grunt-release": "~0.14.0",
-    "grunt-shell": "~1.3.1",
-    "istanbul": "~0.4.5",
-    "mocha": "^3.4.2",
-    "mocha-phantomjs": "~3.5.0",
-    "npm-run-all": "^3.1.1",
-    "phantomjs-prebuilt": "^2.0.0"
+    "babel-cli": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-node6-es6": "^11.2.5",
+    "browserify": "^14.4.0",
+    "chai": "^4.1.2",
+    "commander": "^2.11.0",
+    "core-js": "^2.5.1",
+    "istanbul": "^0.4.5",
+    "jscs": "^3.0.0",
+    "jshint": "^2.9.1",
+    "karma": "^1.7.1",
+    "karma-babel-preprocessor": "^6.0.1",
+    "karma-browserify": "^5.1.1",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-edge-launcher": "^0.4.2",
+    "karma-firefox-launcher": "^1.0.1",
+    "karma-ie-launcher": "^1.0.0",
+    "karma-mocha": "^1.3.0",
+    "karma-mocha-reporter": "^2.2.4",
+    "karma-phantomjs-launcher": "^1.0.4",
+    "karma-safari-launcher": "^1.0.0",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-tap-reporter": "0.0.6",
+    "karma-webpack": "^2.0.4",
+    "mocha": "^3.5.3",
+    "mocha-lcov-reporter": "^1.3.0",
+    "regenerator-runtime": "^0.11.0",
+    "webpack": "^3.6.0",
+    "webpack-merge": "^4.1.0"
   },
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=6"
   },
   "keywords": [
     "JSON",
@@ -68,12 +81,21 @@
     "digital signatures"
   ],
   "scripts": {
-    "lint": "grunt jshint",
-    "jscs": "grunt jscs",
-    "test-all": "npm-run-all --silent test lint jscs",
-    "test-node": "grunt mochaTest",
-    "test-browser": "grunt shell:testBrowser",
-    "test": "grunt mochaTest shell:testBrowser",
-    "coverage": "grunt shell:coverage"
+    "prepublish": "npm run build",
+    "build": "npm run build-webpack && npm run build-node6",
+    "build-webpack": "webpack",
+    "build-node6": "babel --no-babelrc --out-dir dist/node6 --presets=node6-es6 lib/*.js",
+    "test": "npm run test-node",
+    "test-node": "NODE_ENV=test mocha --delay -t 30000 -A -R ${REPORTER:-spec} tests/test.js",
+    "test-karma": "karma start",
+    "coverage": "istanbul cover ./node_modules/.bin/_mocha -- --delay -t 30000 -u exports -R ${REPORTER:-spec} tests/test.js",
+    "coverage-lcov": "istanbul cover ./node_modules/.bin/_mocha --report lcovonly -- --delay -t 30000 -u exports -R ${REPORTER:-spec} tests/test.js",
+    "coverage-report": "istanbul report",
+    "jscs": "jscs lib/*.js tests/*.js",
+    "jshint": "jshint lib/*.js tests/*.js"
+  },
+  "browser": {
+    "lib/index.js": "./lib/jsonld-signatures.js",
+    "crypto": false
   }
 }
